Add allowArbitraryValues to <OwnerPicker />
This change makes the OwnerPicker able to validate that the selected
option is one of the provided values. This should provide a smoother
experience for users, as they now can only enter valid values on the
client side

None of the tests are changed, as this value is passed to MUI, which
is the library containing the validation.

Signed-off-by: kjuulh <contact@kjuulh.io>