Bazaar: Link to the member user catalog entity
When a user is added as a member to a bazaar project the entity ref of
the current user is added to the back end api call.

If the optional user entity ref is available the link for a member will
point to the backstage catalog page for the user. This is instead of the
current "https://github.com/${displayName}".

Signed-off-by: Niklas Aronsson <niklasar@axis.com>