Reset error on success path
If a re-render - or simply the intervalMs parameter - causes a retry that
turns a previously failed operation into a successful one, the error was
previously not reset, resulting in the somewhat odd combination of an
error message and a full, successful, response data object.

This change ensures that the error state is reset every time an operation
succeeds, so the UI state remains consistent.

Signed-off-by: Tomas Aschan <tomasl@spotify.com>