Bazaar plugin: Remove 'Router' as 'RoutedTabs' children
During the 'react-router-dom' upgrades the Bazaar routing broke. This
patch removes the wrapping 'Router' from the 'RouterTabs' children.

Opening the Bazaar plugin would result in the followin error:

"A <Route> is only ever to be used as the child of <Routes> element,
never rendered directly. Please wrap your <Route> in a <Routes>."

Also removed 'react-router-dom' as a peer dependency of the bazaar
plugin.

Signed-off-by: Niklas Aronsson <niklasar@axis.com>
Signed-off-by: Spencer Henry <shenry@qualtrics.com>