Update docker guide for clarity
As happened to a colleague of mine and might happen to others, one might see a "docker" section and skip straight to that since it looks on the surface to be a quick way to check out backstage yourself. In this case that resulted in trying to make all these builds from the main backstage/backstage repo, before even understanding the concept of a Backstage App, which obviously did not work out.

Signed-off-by: Tomás Mota <tomasrebelomota@gmail.com>