Deep-import LightAsync to improve module-import time in tests
`LightAsync` is a `react-syntax-highlighter` that defers the computation
of some fancy syntax-parsing logic.
However, `react-syntax-highlighter` exposes a `Light` variant that does
this computation up-front.

In runtime, tree-shaking ensures that we only import `LightAsync`, but
tree-shaking doesn't happen in tests (because it's slow).

So, deep-import to avoid incurring the syntax-parsing penalty when that
functionality isn't used in the test itself.

Signed-off-by: Mitchell Hentges <mhentges@spotify.com>