Define a simple yet risky Jest resolver
It needs to be injected into Jest twice (`resolver` config option *and*
`createResolver()`) because we want to cover two cases:
1. Determining where a module is, and
2. Defining a module's ID.

Using the public Jest API, only #1 is possible. So, injection is
necessary.

This is able to make things faster in two ways:
1. In `getModuleID`, we no longer try to resolve symlinks and do any FS
work. Instead, we try to generate a unique ID as simply as possible.
2. In `moduleResolver()`, we avoid some extra FS path-resolving work
that Jest does, and instead just call `resolve.sync()`.

This is ~11% faster on my machine (704ms to 622ms) when re-running a
watched test, but I'm not sure how stable it will be on CI. Let's find
out!

Signed-off-by: Mitchell Hentges <mhentges@spotify.com>