Fix "new" ribbon showing on all plugins in marketplace
The original commit 913145399e added a ribbon to show which plugins
were recently added to the marketplace but on the live site all plugins
were showing as 'new'. The problem is due to relying on the git author
date which is incorrect when using shallow depth clones in the build
process.

Fixed by adding new key to plugin marketplace metadata to be parsed.

Fixes: https://github.com/backstage/backstage/issues/13015
Signed-off-by: Calum Lind <calumlind@gmail.com>