feat(catalog-graph): preserve entity graph options + increment max depth
Currently, all configuration options for the `CatalogGraphCard` are lost
at the "View Graph" link at the bottom which opens the full graph view
using the current entity as root entity.
Esp. for `maxDepth` the default was Infinite / no limit.

The change will preserve options used at the `CatalogGraphCard`
(displayed at the entity page) and additionally, increments the
`maxDepth` option by 1 to increase the scope slightly compared to
the graph already seen by the users.

The default for `maxDepth` at `CatalogGraphCard` is 1.

Closes: #14462
Signed-off-by: Patrick Jungermann <Patrick.Jungermann@gmail.com>