introduce error indicator
So that it does not blindly swallow _any_ error (potentially some should be
logged and tracked by other subsystems), an indicator's `onRejected` method will
only resolve errors that it previously rejected. I'm thinking of the case where
the Fetcher needs to reject in order to _intentionally_ short-circuit a batch of
calls to a single cluster, but then the FanOutHandler needs to catch to _avoid_
short-circuiting a fanout across clusters.

Signed-off-by: Jamie Klassen <jklassen@vmware.com>