use node-fetch for k8s resources
The main advantage this has over using the list*CustomObject methods is that it
won't reject non-OK statuses, so we won't have to be so careful about
catching and rethrowing.

Personally I find it a bit clearer to see the actual requests rather than the
client methods with all those empty arguments, and a side benefit is that this
implementation uses the fetch API -- under the covers, @kubernetes/client-node
is using the deprecated request library.

Signed-off-by: Jamie Klassen <jklassen@vmware.com>