Implement useRouteDescriptor
This returns a function on the form (routeRef) => RouteDescriptor where RouteDescriptor contains information about the route ref as it is mounted in the route table. It provides the route ref chain up to the top-level route ref, the parameters for each such route ref, and a matchUrl string which is a regex that will match the full parameterized path up until this route ref.

This is useful to be able to match e.g. the current full path of the browser, with route refs, to see which match.

Signed-off-by: Gustaf Räntilä <g.rantila@gmail.com>