Don't erode webpack error message details
To work around a `react-dev-utils` compatibility problem, errors and
warnings had context stripped before being formatted.

This led to issues in internal Backstage where a build failure's cause
was ambiguous and hard to diagnose.

Since this workaround was needed for `react-dev-utils` v5 (which was
released over a year ago, nice! [1]), I think that it can be removed.

[1] https://github.com/facebook/create-react-app/releases/tag/v5.0.0

Signed-off-by: Mitchell Hentges <mhentges@spotify.com>