GithubDiscoveryProcessor: allow filtering on forks
Forks of repositories containing a catalog yml, might override discovery for
the upstream repository, if both are discovered.

This might not be desired in environments working extensively with forks, as it
can create a lot of confusion and potentially, breakage if fork and upstream end
up with different versions of the catalog yaml.

This changeset provides a way to enable or disable evaluating forked repositories
via Github's API `isFork` attribute.

In detail, this changeset introduces a new config option, namely `allowForks`
(boolean, defaulting to true) and the underlying logic, that will exclude any
forked repository from the list of repositories to be discovered, when set to false.

Signed-off-by: Costas Drogos <costas.drogos@gmail.com>