Tweak CLI admin page example to be compatible with TS linting
Earlier, I assumed that the TS errors were due to Typescript not
understanding that the code was to be run from a "browser" context
rather than "node module" context.

However, simply adding an `export {}` was enough to resolve the
"isolated module"-related warning, and the other type errors were due to
real typing issues which had to be maneuvered around.

Signed-off-by: Mitchell Hentges <mhentges@spotify.com>