can get non-microsoft graph tokens via popup
These tokens will need to be requested on demand with a popup every time: Azure
does allow you to get a refresh token from its `/token` endpoint for a set of
Microsoft Graph scopes and use that refresh token to get access tokens for other
resources, but our session manager isn't quite subtle enough to handle use cases
like that yet.

Note that the function `MicrosoftAuth.expectedClaims` is very nearly a
copy-paste of the `allClaimsForScope` method on `FakeMicrosoftAPI` in the
auth-backend plugin. It would be nice to store this logic in a common location
shared between @backstage/core-app-api and @backstage/plugin-auth-backend, but
it felt hard to justify introducing a whole plugin-auth-common or similar purely
for Azure.

Signed-off-by: Jamie Klassen <jklassen@vmware.com>