feat: add functionality to get scheduled tasks
add new functionality to PluginTaskScheduler to return local and global
tasks. this is to be able to trigger those tasks manually using the
triggerTask functionality when there's no clear way to figure out the
correct id to use as they in many cases come from a plugin.

this is also usable for the upcoming debug plugin, see #9737

Signed-off-by: Heikki Hellgren <heikki.hellgren@op.fi>