Pass configs from build:backend to app builds. Fixes #15274
When using `build:backend` with `--config` options, they're
not passed down to the packaging of the FE app.
Thus, that's picking up the default configs, which includes
the `app-config.local.yaml`, but it's also missing the
production yaml, if requested.

Pass the given `configPaths` around for the backend build like
the frontend build already does.

Signed-off-by: Axel Hecht <axel@pike.org>