feat: add documented component to uffizzi environment
So far, none of the components at the uffizzi environment have techdocs.
Hence, there are no docs available within this environment
and related features are not covered.

This change will add an example from techdocs-backend plugin
to the deployed examples.

It was possible to register this catalog file manually,
however the build of the docs fail due to missing Docker.

To allow the generation of docs, we use the local generator
(no docker in docker) and install required dependencies.

As we don't want all dependencies for all (default) extensions
to be installed at this environment, the original example at
techdocs-backend plugin was copied and reduced
(PlantUML extension example was removed).

Signed-off-by: Patrick Jungermann <Patrick.Jungermann@gmail.com>