feat(github): support commit hashes at GithubUrlReader.readTree/search
Support commit hashes additionally to branch names at the `GithubUrlReader`'s `readTree` and `search`.

This is achieved by using the "combined commit status" API instead of the branch API
which works with commit hashes and branch names.
When used with branch names, it will return the result for the HEAD of the branch
which is what the branch API was used for before.
We are not really interested in commit statuses, though, and can ignore them.

The "combined commit status" API support branch names with slashes as well, like `branch/name`.

Additionally, this will reduce the number of API calls from 2 to 1 for retrieving the "repo details"
for all cases besides when the default branch has to be resolved and used
(e.g., repo URL without any branch or commit hash).

Relates-to: PR #16721
Signed-off-by: Patrick Jungermann <Patrick.Jungermann@gmail.com>