Omit ManagedFields from the Kubernetes resource YAML display. (#17497)
* Omit ManagedFields from the Kubernetes resource YAML display.

The default for the Kubernetes CLI tooling is to omit the managed fields
by default, since these are only present for internal machinery. For
the Kubernetes resource YAML display, the managed fields take up a ton
of space, but are never useful to operators who are interested in the
details of their Kubernetes resources.

This fixes #14177.

Signed-off-by: James Peach <jpeach@apache.org>

* Add a control to toggle whether the managed fields display.

So that managed fields can be used for debugging purposes, add a switch
to toggle allow managed fields to display. This is off by default.

Signed-off-by: James Peach <jpeach@apache.org>

---------

Signed-off-by: James Peach <jpeach@apache.org>