feature(plugin-search-react): Extend InputProps on <SearchBar/>
In the current implementation of <SearchBar/> it is possible to
overwrite the default <TextField/> InputProps by passing a prop of the
same name. But sometimes is necessary to extend only, for example when
one is interested on changing the clear button only.
The clear button is supplied in the endAdornment, part of the
InputProps, but at the same time the startAdornment is also on it.
To be able to replace the clear button implementation without affecting
the rest, this patch proposes a change where `InputProps` is exposed in
the same manner as `inputProps` (notice the change from capital I to i).
Where the received object may overwrite keys using the spread operator.

Signed-off-by: Renan Mendes Carvalho <aitherios@gmail.com>