docs: Add error handling instructions for 'No token available' error
This commit adds instructions to the Backstage documentation on how to handle the `No token available for host: github.com, with owner , and repo from-backstage' error`. The instructions guide the user to restart Backstage using `Control-C` and `yarn dev` commands in the terminal, followed by clicking on the **START OVER** button.

Signed-off-by: sammy. <samuelumoren365@gmail.com>
Signed-off-by: Samuel <samuelumoren365@gmail.com>