incident plugin, backend and frontend
Author two new plugins designed to work together:

- plugins/incident => to display associated incident.io incidents in the
  entity dashboard page, under an overview card.
- plugins/incident-backend => syncing catalog data from Backstage into
  incident.io using the scheduled tasks system.

Users of this plugin are expected to sync Backstage data into
incident.io in some way – either using the backend plugin in this
commit, or the catalog-importer – and create incident custom fields such
as "Affected components" which are powered by the synced catalog data.

Then the frontend plugin is configured to point at those fields so it
can find any associated incidents.