✨Make incremental event handler to be both explicity and async
The point of the event hook is to map incoming changes into catalog
deltas. However, this is not always a synchronous process, and might
require accessing other resources to get the actual delta.

This just converts the hook into an async function and awaits it on
the other end. This also cleans up the API a little bit by not relying
on `undefined` to indicate that the event was ignored, but by
explicitly forcing the event handler to indicate that the event is to
be ignore by returning { type: "ignored" }.

Since ignoring events is a perfectly acceptable outcome, the warning
if the event was dropped is just ignored. If there is an error, it
should be propagated as an exception.

This is purely an aesthetic change designed to communicate via usage
how to use the API.

Signed-off-by: Charles Lowell <cowboyd@frontside.com>