fix(deps): fix SNYK-JS-FASTXMLPARSER-5668858
Fixes SNYK-JS-FASTXMLPARSER-5668858
by upgrading aws-sdk dependencies to
their latest version.

The upgrade of `fast-xml-parser` to
version 4.2.4 (or higher) was included
in release v3.347.1 of the aws-sdk.

Latest release is v3.350.0, however,
not for all packages.

The previous version's requirement would have
allowed to use a newer version of the aws-sdk
to fix it locally to a backstage project setup.

This change will enforce a version including the fix.

Closes: #18164
Signed-off-by: Patrick Jungermann <Patrick.Jungermann@gmail.com>