feat: allow processing placeholders in same directory
This change allows to resolve YAML placeholders that are located in the same
directory as the source is. This enables to split entity files, especially
very long templates, to multiple parts inside the same directory.

For example in #15241 the $yaml includes can now be done now in local
filesystem instead fetching them from the remote. Only limitation is
that the included files must be in the same directory (or subdirectory)
as the file including those is.

Signed-off-by: Heikki Hellgren <heikki.hellgren@op.fi>