fix(bitbucketCloud,events): handle missing slug in event payload
The current handling for `repo:push` events assumed that
the repo slug is provided as part of the event.

However, this is not the case which led to a few issues:

- repo slug filter was always positive
  as it was checked against `undefined`
- the `repo:` filter at the code search query was never applied
  which led to the same result as at the full refresh
  - all additionally discovered catalog files were categorized as "added"
  - this may not have caused additional API requests (not verified)

This change will enhance the event by adding the missing slug,
extracted from the `full_name`
(format: `{workspace-slug}/{repo-slug}`).

Signed-off-by: Patrick Jungermann <Patrick.Jungermann@gmail.com>