fix(confluence-to-markdown): update config struture
This change is being implemented so there is not a conflict between
confluence-to-markdown and the confluence search plugin. Both items
make use of `confluence.auth`. The confluence-to-markdown expects
`auth` to be a string where as within the search plugin it is an
object. Similarly both items expect a username and password or a
token. In order to elimiate the duplicate code, this action is being
updated to make use of the `auth` object and all items moving under
this object.

Signed-off-by: Jonathan Sundquist <jsundquist@gmail.com>
Signed-off-by: Gaurav Pandey <gpandey1@volvocars.com>