Update v1.16.0.md
The "Mailchimp" link seems to have stoped working. Suggestion to replace the link in the latest release note - as well as the upcoming release notes - with the link used for the newsletter in other place. Let me know if there is a template somewhere where I can apply this change as well.

Signed-off-by: Philipp Hugenroth <tudi2d@users.noreply.github.com>