fix(auth/microsoft): fix missing profile photo and access token for foreign scopes
This change re-adds the loading of the logged-in user's profile picture
as well as the "skip profile" logic for foreign scopes
which got lost while migrating the Microsoft auth provider to the new module.

Additionally, it uses the 96x96 size instead of 48x48.
The image is used on the profile card on the settings page (`/settings`)
rendered as 96x96 and was blurry due to the upscaling.

Replaces the `types.d.ts` file with type declarations with the use of
`@types/passport-microsoft`.

Closes: #20677
Closes: #20699
Signed-off-by: Patrick Jungermann <Patrick.Jungermann@gmail.com>
Signed-off-by: Tauhid <tauhidanjum@gmail.com>