Allow globby negative matches for copyWithoutTemplating
The option copyWithoutTemplating (and depracted copyWithoutRender) would
require listing all files in a path to include even if there was only
one or two files that should still be templated while skipping all
others. This works fine when just one or two files or paths that need to
be copied without being templated, however when there are many files and
only one or two should be templated, it becomes more error prone to try
and maintain a complete list and ensure others will update it correctly.

While it's possible to use separate paths to avoid this issue, it
results in subsequently needing to move files around, or use separate
template calls, this makes it more difficult to maintain a template
logically.

The existing template code already makes use of 'globby', which supports
negative matching to explicitly exclude individual paths from other
wildcard matches. The existing code would iterate over each entry
calling globby separately for each pattern passed in. This would skip
globby's ability to have negative matches remove previously matched
paths.

This change switches to pass the entire list of patterns to globby in a
single call, leaving it up to globby to parse and return the entire set
in a single operation.

Signed-off-by: Darragh Bailey <dbailey@g-p.com>
Signed-off-by: Bogdan Nechyporenko <bnechyporenko@bol.com>