[SupportButton] allow custom icon for SupportItem
Allow the usage of non-predefined icons for SupportItem list.

Part of the reason SupportItems can be supplied to modify the
SupportButton entries is to provide links to the users support tool of
choice. Allowing the usage of custom icons and not just the pre-defined
set will make it possible to make this experience nicer, by using icons
the users may more readily associate with that tool.

Signed-off-by: Rickard Dybeck <dybeck@spotify.com>