feat(tech-insights): add plugin for new backend system
A new backend plugin for the tech-insights backend
was added and exported as `default`
as well as extension points to be used by modules.

The plugin can be used like

```ts title="packages/backend/src/index.ts"
backend.add(import('@backstage/plugin-tech-insights-backend'));
```

Fact retrievers will be added
(built-in or through the extension point),
but only registered if the user adds a configuration for it
(cadence, etc.).

Signed-off-by: Patrick Jungermann <Patrick.Jungermann@gmail.com>