Gerrit readUrl: Use gitilies if available
Current implementation of the Gerrit readUrl only supports fetching
files from the tip of a branch. There are plenty of usecases where
we need to reference other files in a Backstage context.

Just as previously was done with the "readTree" implementation this
makes readUrl fetch the file content from Gitiles (if configured).

The file content of a Gitiles url can be fetch by appending "?format=TEXT"
to the url making it easy to support any type of urls.

Using the Gerrit REST API to make the readUrl more flexible would be
complicated. First the Gitiles url must be parsed to determine the
type of git reference it points to. Then depending on the type we
either need to query the "changes" or "projects" API. To support "tags"
we would first have to query the "tag api" to get the revision the
tag is pointing to and from there look up the content.

Signed-off-by: Niklas Aronsson <niklasar@axis.com>