Scaffolder-backend-gerrit: Set branches to defaultBranch
When a project is created in `publish:gerrit` the default branch needs
to be provided to the Gerrit API that creates the project. If the
branches attribute is not provided, the default branch for the project
will be set to whatever is configured as default on the Gerrit host.

Reference: https://gerrit-review.googlesource.com/Documentation/rest-api-projects.html#project-input
Signed-off-by: Martin Wallgren <martinwa@axis.com>