beps/0003: add TODOs for issuing service tokens
Couple of concerns around how service tokens are issued that came up. First one is whether we should require callers to specify which plugin they want to talk to - as this information can be quite valuable to the implementation. Second is how we make it hard to forget to forward credentials, since the existing interface doesn't really do that.

Signed-off-by: Patrik Oldsberg <poldsberg@gmail.com>