Avoid flicker when refreshing page of new frontend system
Due to something (?) async, when refreshing a page using the new frontend system,
there's a brief flicker as the `React.Suspense` fallback is shown (which was just
the text `Loading...`). Instead, let's show a `<Progress />` - it seems to make
the flicker disappear when reloads are fast.

Either way, the UX feels mildly nicer to see a loading bar on the top of the
screen instead of the text `Loading...`.
Unfortunately, one downside here is that, if running on a slower network, then
the loading bar is shown in two different colours: once unthemed, then again
themed.

Signed-off-by: Mitchell Hentges <mhentges@spotify.com>