integration: use href from git-url-parse when resolving file urls
When parsing a URL with a trailing slash,
git-url-parse drops the trailing slash from the
href and filepath properties on the result, which
is inconsistent with the behavior of the native
URL constructor. Since before we were using the
filepath property from git-url-parse to determine
how much to truncate the pathname from the URL
constructor, we were off-by-one for URLs with a
trailing slash.

After this change, we pass the href property from
git-url-parse to the URL constructor, such that
the amount to truncate the pathname and the
pathname itself are ultimately coming from the
same source.

Signed-off-by: MT Lewis <mtlewis@users.noreply.github.com>