Fix absolute path resolution for Github
The logic for constructing the `repoRootPath` was incorrectly leaving the last character of the preceding directory when trimming the 'filepath' from the 'base' URL. This was due to an off-by-one error where the length of `filepath` did not account for the leading slash. The additional `-1` subtraction ensures that both the `filepath` and its leading slash are completely removed, resulting in the correct repository root path.

Previously, if given a `base` of `https://github.com/backstage/backstage/tree/master/packages/cli/` and an absolute path of `/package.json` the URL would resolve to `https://github.com/backstage/backstage/tree/master/p/package.json`.

Signed-off-by: Brian Hudson <brian.r.hudson@gmail.com>