Propose changes to better support external notifications
While it makes sense to send notifications to users (or broadcast) in many scenarios, its possible that external integrations via NotificationProcessor may not have a 1:1 relationship with users.

For example, while the current interface is adequate for sending Slack DMs to users or teams (via the team expansion code in the notifications-backend) it doesn't facilitate a way to choose a different destination such as a Slack channel or Email mailing list.

We could add another recipient type of `external` that would accept some arbitrary value that processors can choose to support. For example a value for `externalRef` may be something like `slack:C03SABTG6RX` or `slack:#abc` or `email:list@example.com`.

This would also change the interface for recipients to accept an array so that multiple types of recipients could be specified.

This would still allow for direct user notifications if desired, as well as separate external notifications sent for the same `SendNotificationRequest`.

One challenge with this approach might be if you want a notification to go to individual users using the WebProcessor and only 1 notification via the SlackProcessor as there is not a way for the processor to know whether it should act on all notifications it receives or not.

Signed-off-by: Adam Kunicki <kunickiaj@gmail.com>