feat: run mkdocs via Docker Engine API
The system may not have the Docker CLI installed while still having
access to an implementation of the Docker Engine API. For example, if
the daemon is running in a VM with shared folders, or when another
container runtime is being used.

This changeset modifies techdocs-cli to interact with Docker via the
Engine API instead of the Docker CLI. The local Docker Engine UDS is
used by default, but another location can be specified with the
`DOCKER_HOST` environment variable.

Fixes: #12028
Signed-off-by: Terin Stock <terinjokes@gmail.com>