(#25103) Add oidc-proxy backend auth provider
Previously, the various [proxy based auth providers][1] are all tied to
a specific identity aware proxy.  For example, the gcp-iap provider is
tightly coupled to the GCP identity aware proxy.

This is a problem for similar architectures such as Holos Platform,
which uses Istio with ExtAuthZ to provide an oidc id token in the
`x-oidc-id-token` header.  The existing providers cannot be easily
re-used for such an integration even though the only major difference is
the OIDC issuer url and the header used to carry the id token.

This patch solves the problem by implementing a generally useful
`auth-backend-module-oidc-proxy-provider` with a provider id of
`'oidc-proxy'`.

This backend auth provider is implemented as a nearly perfect copy of
the existing gcp-iap provider, which the following notable exceptions.

 1. Integrators may configure the oidc issuer with oidc-proxy using the
    `iss` Config field.
 2. Integrators may configure the header the id token is expected at
    using the `oidcIdTokenHeader` Config field.

Both of these settings are hard-coded to Google specific values through
the use of the `google-auth-library` in the case of the `gcp-iap`
provider.

Otherwise, this `oidc-proxy` provider strives to stay true to the spirit
and behavior of the `gcp-iap` provider in an effort to reduce the
cognitive load of the core maintainers and integrators.

[1]: https://github.com/backstage/backstage/blob/237d7e832541fb990bdee0bd5af1840c3b790e1b/plugins/auth-node/src/proxy/types.ts#L35-L40

Signed-off-by: Jeff McCune <jeff@openinfrastructure.co>