Update `patchMkdocsYmlPreBuild`
Update the `patchMkdocsYmlPreBuild` function to modify `repo_url` and
`edit_uri` independently of each other. Prior to this change both
`repo_url` and `edit_uri` would have to be missing for either to be
generated which caused an issue in the situation where `repo_url`
is provided in the mkdocs.yml and `edit_uri` is not because the default
`edit_uri` value from mkdocs would be used instead. This value uses `master` as
the branch in the link which breaks the edit link for repos that don't use
`master` as the default branch.

Signed-off-by: Isabel Tomb <isabelgtomb@gmail.com>