Update test
Update the test that checks that the existing edit_uri value in a
mkdocs.yml won't be overridden by `patchMkdocsYmlPreBuild` so that it
checks the value of 'edit_uri' was not updated rather than just checking
for the existence of the substring. This change is needed because the
substring will exist in the file but as the `repo_url` value which is
expected after the changes in the previous commits.

Signed-off-by: Isabel Tomb <isabelgtomb@gmail.com>