feat: allow signing git commits from scaffolder
adds support to sign git commits with configured PGP key in scaffolder
actions. configuration can be done either by integration or by using the
default signing key in the scaffolder config. note that this
pgp-plugin is used for signing and that it is limited to using RSA keys
and signatures made with SHA1 hashing algorithm.

closes #25934

Signed-off-by: Heikki Hellgren <heikki.hellgren@op.fi>