yarn-plugin: add additional check during pack to confirm that all backstage versions have been replaced
Adds an additional check during pack which walks all the
dependencies in the manifest, and verifies that
there are no remaining `backstage:` versions.
I'm not aware of any realistic cases where the
code that does the replacement would be expected
to fail, but since a backstage:^ package in the
manifest will break the release, it seems prudent
to double check and fail fast to allow us to catch
unexpected results.

Signed-off-by: MT Lewis <mtlewis@users.noreply.github.com>