implement as a ThrottleService instead
This avoids the confusion of different fetch versions on different platforms (node vs browser vs polymorphic), by implementing the throttler as a concept of its own that does not have to know about how requests and responses work.

In all honesty, in its current form, it doesn't even have to be tied to fetch specifically; it could throttle ANY function. But let's not over-generalize it so it gets misused.

Signed-off-by: Fredrik Adelöw <freben@gmail.com>