yarn-plugin: move Backstage version to parameters in bound descriptors
The previous version of this code transformed
the version range "backstage:^" to
"backstage:<version-from-backstage.json>" when
writing entries in yarn.lock. This does seem to be
something that the yarn plugin API supports, but
it turns out to cause issues in some third-party
tools which assume that version ranges will be
consistent between package.json and yarn.lock.

To fix this issue, we're now leveraging the
descriptor binding feature in Backstage (which I
didn't know about before) to append a descriptor
parameter containing the Backstage version. This
keeps the selector of the descriptors consistent
between package.json and yarn.lock entries, which
will hopefully mean that this plugin no longer
breaks other tools.

cf. https://github.com/snyk/nodejs-lockfile-parser/issues/256

Signed-off-by: MT Lewis <mtlewis@users.noreply.github.com>